requirements.txt
flask
openai>=1.40.0
python-dotenv
